refactor: extract pending setup first steps route-local component
